add a launcher utility used to run zeek on windows
This launcher is responsible for constructing the required zeek
environment variables and running the zeek binary embedded in the
zdeps/zeek directory in Brim.

It's the windows equivalent of the small shell script for linux/darwin
that execs zeek, but more complicated due the cygwin path translation
and the extra hoops needed to ensure zeek is terminated when this
process is.