transition to use zeekrunner (#23)
The script (mac/linux) or executable (windows) that runs Zeek, including creating required environment variables & command line options, is now zeekrunner or zeekrunner.exe . The command line options used to run Zeek are now in this repo instead of in the zqd source. The job object based process termination for windows is removed, as it is moved into zqd; this will make it easier to eventually replace the windows zeek runner with a script or batch file.